refactor: 그룹 컴포넌트 디렉토리 재구성 및 라우팅 구조 개선
