--- src/components/group/GroupDetail.jsx
@@ -1,3 +1,0 @@
-export default function GroupDetail() {
-  return <div>그룹 정보</div>;
-}
