+++ src/components/group/detail/GroupDetail.jsx
@@ -1,0 +1,12 @@
+import groupImg from "@/assets/Image 4.svg";
+
+export default function GroupDetail() {
+  return (
+    <div>
+      <div>
+        {" "}
+        <img src={groupImg} alt="그룹이미지" />
+      </div>
+    </div>
+  );
+}
