+++ src/components/group/main/GroupCard.module.css
@@ -1,3 +1,8 @@
+.groupLink {
+  color: inherit;
+  text-decoration: none;
+}
+
 .groupCard {
   display: flex;
   flex-direction: column;
