+++ src/components/group/shared/SearchBar.jsx
