+++ src/pages/Group.jsx
@@ -1,9 +1,9 @@
 import Header from "@components/all/Header";
-import GroupDetail from "@components/group/GroupDetail";
-import SearchBar from "@/components/group/SearchBar";
-import CardList from "@/components/group/CardList";
-import LoadMoreButton from "@/components/group/LoadMoreButton";
-import PageLayout from "../components/group/PageLayout";
+import GroupDetail from "@components/group/detail/GroupDetail";
+import SearchBar from "@/components/group/shared/SearchBar";
+import CardList from "@/components/group/shared/CardList";
+import LoadMoreButton from "@/components/group/shared/LoadMoreButton";
+import PageLayout from "../components/group/shared/PageLayout";
 
 export default function Group() {
   return (
