+++ src/pages/GroupRegister.jsx
@@ -1,5 +1,5 @@
 import Header from "@components/all/Header";
-import GroupSetup from "@components/group/GroupSetup";
+import GroupSetup from "@components/group/shared/GroupSetup";
 
 export default function GroupRegister() {
   return (
